backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
aiogram
pydantic-settings
alembic
python-dotenv
